Facade is the class for simplifying the work of the teacher. When a teacher wants to gather the class and make an attendance check then dismissing it, or dismissing it and making n attendancecheck or something else, it uses facade class' instance. The facade class instance abstracts the teacher from pinging the students or gathering them.
Teacher class have only two methods. These methods do not deal wih the students.
İnstead of dealing with the students, it uses the facade.

Student class' attendancePing method replies the call from the facade class instance.
The constructor gives a value of "0" to the member named "dismissTime".

Facade class instance looks for the "dismissTime" value.
If it is "0" than it records student as "absent". else, the student is there.